Create the v2 router that uses channels internally
The v2 router is slightly more complicated than the v1 router becauses
it uses channels to pass messages to different threads in the
application.

There is a single async loop that is responsible for processing all the
`Command` events that it receives. This includes user connections, user
disconnections, and messages sent to a specific user.

The application now has state associated with it. The main peice of
state that gets stored by the app is the transmitting end of the main
async command loop so that it can be passed to all routes that need to
issue commands.