Add v3 version of the APP
The v3 version of the app is identical to the v2 version with the
addition of Kafka support. The v3 verison of the app listens for new
messages on the `v3_messages` topic as another input source for users
messages.

Because the application uses channels to pass messages around internally
it was easy to add support for kafka by sending the stream of messages
from kafka to the main async loop.

In order to setup kafka more easily for end users a `docker-compose.yml`
file was added to get `kafka` and `zookeeper` up and running with a
simple:

```
docker-compose up -d --build
```