Add endpoint to change a report status
This endpoint recieves a report status update JSON message from the
client and forwards that message to a Kafka producer. Since the app is
already listening for kafka messages I wanted to get some hands on
experience working with a producer. Turns out it's really simple :^)

In theory it would be easier and more efficient to send the  message
directly to the main loop that's already processesing tasks
(`handle_app_events`), but where would the fun (and learning) be in
that. Now we can better simulate what it would be like to receive
messages from the Kafka topic.