update css to display flex items at starting from the top of the page
For the demo app it's a little nicer to display all the items at the top
of the page. This allows the app to be displayed on the right of the
page, the browser console to be open on the left, and a terminal window
open accross the bottom where I can run the `update_report_status.sh`
script to transition reports to new statuses.